fix diffusion samplers
