+++ src/opensynth/models/energydiff/sampler.py
@@ -6,8 +6,6 @@ import math
 
 from tqdm.auto import tqdm
 
-from .diffusion import GaussianDiffusion1D, ModelMeanType, ModelVarianceType
-
 " *** check the bottom of this script for sample functions      *** "
 " --- below is mostly DPM-Solver implementation and adaptation  ---"
 class NoiseScheduleVP:
@@ -1191,198 +1189,3 @@ def expand_dims(v, dims):
     """
     return v[(...,) + (None,)*(dims - 1)]
 
-class DPMSolverSampler(object):
-    def __init__(self, model: GaussianDiffusion1D, **kwargs):
-        super().__init__()
-        self.model = model
-        device = next(model.parameters()).device
-        to_torch = lambda x: x.clone().detach().to(torch.float32).to(device)
-        self.register_buffer('alphas_cumprod', to_torch(model.alpha_cumprod))
-
-    def register_buffer(self, name, attr):
-        if type(attr) == torch.Tensor:
-            if attr.device != torch.device("cuda"):
-                attr = attr.to(torch.device("cuda"))
-        setattr(self, name, attr)
-
-    @torch.no_grad()
-    def sample(self,
-               S, # steps
-               batch_size, # N
-               shape, # (C, L)
-               conditioning=None, # (N, *) or broadcastable (1, *)
-               callback=None,
-               normals_sequence=None,
-               img_callback=None,
-               quantize_x0=False,
-               eta=0.,
-               mask=None,
-               x0=None,
-               temperature=1.,
-               noise_dropout=0.,
-               score_corrector=None,
-               corrector_kwargs=None,
-               verbose=True,
-               x_T=None,
-               log_every_t=100,
-               cfg_scale=1.,
-               unconditional_conditioning=None, # the unconditional-class tensor
-               # this has to come in the same format as the conditioning, # e.g. as encoded tokens, ...
-               **kwargs
-               ):
-        if conditioning is not None:
-            if isinstance(conditioning, dict):
-                cbs = conditioning[list(conditioning.keys())[0]].shape[0]
-                if cbs != batch_size and cbs != 1:
-                    print(f"Warning: Got {cbs} conditionings but batch-size is {batch_size}")
-            else:
-                if conditioning.shape[0] != batch_size and conditioning.shape[0] != 1:
-                    print(f"Warning: Got {conditioning.shape[0]} conditionings but batch-size is {batch_size}")
-
-        # sampling
-        (C, L) = shape
-        size = (batch_size, C, L)
-
-        # print(f'Data shape for DPM-Solver sampling is {size}, sampling steps {S}')
-
-        device = next(self.model.parameters()).device
-        if x_T is None:
-            img = torch.randn(size, device=device)
-        else:
-            img = x_T
-
-        ns = NoiseScheduleVP('discrete', alphas_cumprod=self.alphas_cumprod)
-
-        # original model function
-        def apply_model(x, t, c=None):
-            out = self.model.model.forward(x, t)
-            if self.model.model_variance_type == ModelVarianceType.LEARNED_RANGE: # GaussianDiffusion1D.model_variance_type
-                out = torch.split(out, out.shape[1] // 2, dim=1)[0] # discard the learned variance
-            return out
-        # model mean type
-        if self.model.model_mean_type == ModelMeanType.NOISE:
-            model_type = 'noise'
-        elif self.model.model_mean_type == ModelMeanType.X_START:
-            model_type = 'x_start'
-        elif self.model.model_mean_type == ModelMeanType.V:
-            model_type = 'v'
-        else:
-            raise ValueError(f"Unknown model mean type {self.model.model_mean_type}")
-        model_fn = model_wrapper(
-            apply_model,
-            ns,
-            model_type=model_type,
-            guidance_type="classifier-free",
-            condition=conditioning,
-            unconditional_condition=unconditional_conditioning,
-            cfg_scale=cfg_scale,
-        )
-
-        dpm_solver = DPM_Solver(model_fn, ns, predict_x0=True, thresholding=False)
-        x = dpm_solver.sample(img, steps=S, skip_type="time_uniform", method="multistep", order=3, lower_order_final=True)
-
-        return x.to(device), None
-    
-# two functions we use to sample. 
-def ancestral_sample(
-    num_sample: int,
-    batch_size: int,
-    data_sequence_length: int,
-    data_feature_dim: int, # 1 for univariate, 2 for bivariate
-    # cond: torch.Tensor|None=None,
-    # cfg_scale: float = 1.,
-    diffusion_model: GaussianDiffusion1D,
-) -> torch.Tensor:
-    """ sample from either given diffusion model. 
-    
-    Arguments:
-        - num_sample: int, number of samples to generate
-        - batch_size: int, batch size for sampling
-        - cond: shape [batch_size, channel, 1], condition for sampling
-    """
-    # process cond
-    # if cond is not None:
-    #     if model is not None:
-    #         _device = next(model.parameters()).device
-    #         cond = cond.to(_device)
-    #     else:
-    #         cond = cond.to(trainer.device)
-    
-    # process batch size split
-    if num_sample < batch_size:
-        list_batch_size = [num_sample]
-    else:
-        list_batch_size = [batch_size] * (num_sample // batch_size)
-        if num_sample % batch_size != 0:
-            list_batch_size.append(num_sample % batch_size)
-    
-    # sample
-    list_sample = []
-    for idx, batch_size in enumerate(list_batch_size):
-        print(f'sampling batch {idx+1}/{len(list_batch_size)}, batch size {batch_size}. ')
-        # cond = cond[:batch_size] if cond is not None else None
-        # model_kwargs = {
-        #     'c': cond,
-        #     'cfg_scale': cfg_scale,
-        # }
-        # 
-        # if model is None:
-            # trainer.ema.ema_model.half()
-            # with trainer.accelerator.autocast():
-        #     sample_batch = trainer.ema.ema_model.sample(batch_size=batch_size, clip_denoised=True, model_kwargs=model_kwargs).float()
-        # else:
-        #     sample_batch = model.sample(batch_size=batch_size, clip_denoised=True, model_kwargs=model_kwargs)
-        sample_batch = diffusion_model.p_sample_loop(
-            shape=(batch_size, data_sequence_length, data_feature_dim),
-            noise=None,
-            clip_denoised=False, # True requires data scaling to (-1,1)
-        )
-
-        list_sample.append(sample_batch)
-    all_sample = torch.cat(list_sample, dim=0)
-    
-    return all_sample # shape [num_sample, data_sequence_length, data_feature_dim]
-
-
-def dpm_solver_sample(
-    sampler: DPMSolverSampler,
-    total_num_sample: int,
-    batch_size: int,
-    step: int,
-    shape: tuple[int, int],
-    conditioning: torch.Tensor|None,
-    cfg_scale: float,
-    accelerator = None,
-    clip_denoised: bool = True,
-) -> torch.Tensor:
-    if accelerator is not None:
-        num_sample = total_num_sample // accelerator.num_processes # for this process
-    else:
-        num_sample = total_num_sample
-    if num_sample < batch_size:
-        list_batch_size = [num_sample]
-    else:
-        list_batch_size = [batch_size] * (num_sample // batch_size)
-        if num_sample % batch_size != 0:
-            list_batch_size.append(num_sample % batch_size)
-    
-    list_sample = []
-    for idx, batch_size in enumerate(list_batch_size):
-        print(f'sampling batch {idx+1}/{len(list_batch_size)}, batch size {batch_size}. ')
-        sample_batch, _ = sampler.sample(
-            S = step,
-            batch_size = batch_size,
-            shape = shape,
-            conditioning = conditioning,
-            cfg_scale = cfg_scale,
-        )
-        list_sample.append(sample_batch)
-        
-    all_sample = torch.cat(list_sample, dim=0)
-    if accelerator is not None:
-        gathered_all_sample = accelerator.gather(all_sample)
-        return gathered_all_sample
-    
-    if clip_denoised:
-        all_sample = torch.clamp(all_sample, -1, 1)
-    return all_sample
